ADDED legacy getters getFirstCorner

+++ legacy/src/cliffscalculator.cpp
@@ -7,7 +7,7 @@
 #define ZERO_INCLIN2 0
 
 // inline specialisation
-template<typename T> T CliffsCalculator::delta1At(ListSwitchable<T>& This,int position){
+template<typename T> T CliffsCalculator::delta1At(ListSwitchable<T>& This,size_t position){
 
     if (position<This.size()){
 
@@ -45,9 +45,9 @@ template<typename T> T CliffsCalculator::delta1At(ListSwitchable<T>& This,int po
     // never happens
     return T(0,0);
 }
-template dpoint CliffsCalculator::delta1At(ListSwitchable<dpoint>& This,int position);
+template dpoint CliffsCalculator::delta1At(ListSwitchable<dpoint>& This,size_t position);
 
-template <typename T> qreal CliffsCalculator::inclin1At(ListSwitchable<T>& This,int i){
+template <typename T> qreal CliffsCalculator::inclin1At(ListSwitchable<T>& This,size_t i){
     T d1 = CliffsCalculator::delta1At(This,i);
     //if (d1.x()==0) {
     //    return 0;
@@ -55,14 +55,14 @@ template <typename T> qreal CliffsCalculator::inclin1At(ListSwitchable<T>& This,
     return d1.y()/d1.x();
     //}
 }
-template qreal CliffsCalculator::inclin1At(ListSwitchable<dpoint>& This,int i);
+template qreal CliffsCalculator::inclin1At(ListSwitchable<dpoint>& This,size_t i);
 
 // Warning: this is looking back
-template <typename T> bool CliffsCalculator::deriv1Crossed(ListSwitchable<T>& This,int pos){
-    if(pos < 1)
+template <typename T> bool CliffsCalculator::deriv1Crossed(ListSwitchable<T>& This,size_t pos){
+    if(pos < size_t(1))
         return false;
 
-    qreal d1 = CliffsCalculator::inclin1At(This,pos-1);
+    qreal d1 = CliffsCalculator::inclin1At(This,pos-size_t(1));
     qreal d2 = CliffsCalculator::inclin1At(This,pos);
 
     // This will get the first corner but might be catching too early: d1=0,d2=0
@@ -71,4 +71,4 @@ template <typename T> bool CliffsCalculator::deriv1Crossed(ListSwitchable<T>& Th
     }
     return false;
 }
-template bool CliffsCalculator::deriv1Crossed(ListSwitchable<dpoint>& This,int pos);
+template bool CliffsCalculator::deriv1Crossed(ListSwitchable<dpoint>& This,size_t pos);
